Ledger branch mixin
